ConsumerStateConverter ready.

+++ src/main/java/com/epsm/electricPowerSystemDispatcher/repository/SavedConsumerStateDao.java
@@ -5,6 +5,6 @@ import java.util.List;
 import com.epsm.electricPowerSystemDispatcher.model.domain.SavedConsumerState;
 
 public interface SavedConsumerStateDao {
-	public List<SavedConsumerState> getStatesByNumber(int consumerNumber);
+	public List<SavedConsumerState> getStatesByNumber(long consumerNumber);
 	public void saveState(SavedConsumerState state);
 }
